fix(livecchat): Merge livechat lib into main SDK
Includes lots of fixes to driver and tests
