+++ src/index.ts
@@ -1,8 +1,10 @@
+import * as ddp from './lib/ddp'
 import * as driver from './lib/driver'
 import * as methodCache from './lib/methodCache'
 import * as api from './lib/api'
 import * as settings from './lib/settings'
 export {
+  ddp,
   driver,
   methodCache,
   api,
