--- src/livechat/lib/settings.js
@@ -1,6 +1,0 @@
-"use strict";
-exports.__esModule = true;
-// Livechat settings
-exports.token = process.env.LIVECHAT_TOKEN || '';
-exports.deparmentId = process.env.LIVECHAT_DEPARTMENT || '';
-exports.rid = process.env.LIVECHAT_ROOM || '';
