--- src/livechat/lib/settings.ts
@@ -1,4 +1,0 @@
-// Livechat settings
-export let token = process.env.LIVECHAT_TOKEN || ''
-export let deparmentId = process.env.LIVECHAT_DEPARTMENT || ''
-export let rid = process.env.LIVECHAT_ROOM || ''
