+++ src/utils/livechat/config.ts
@@ -1,6 +1,6 @@
-import * as api from './lib/api'
-import { silence } from '../lib/log'
-import { mockVisitor } from './lib/mock'
+import * as api from '../../lib/api'
+import { silence } from '../../lib/log'
+import { mockVisitor } from '../config'
 
 silence()
 
@@ -9,7 +9,7 @@ const { token } = mockVisitor.visitor
 async function config () {
   console.log(`
 
-Get deafult Livechat Config \`api.livechat.config()\`:
+Get default Livechat Config \`api.livechat.config()\`:
 ${JSON.stringify(await api.livechat.config(), null, '\t')}
 
 Get Livechat Config with Token \`api.livechat.config({ token })\`:
